% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{add_custom_footer}
\alias{add_custom_footer}
\title{Title Add custom footer to a table}
\usage{
add_custom_footer(ft, footer_text)
}
\arguments{
\item{ft}{A flextable object created by create_table function}

\item{footer_text}{Footer text to be added}
}
\description{
Title Add custom footer to a table
}
