% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{bind_conf_interval}
\alias{bind_conf_interval}
\title{Title Binds mean and 95% CI}
\usage{
bind_conf_interval(V1, ci_lower, ci_upper)
}
\arguments{
\item{V1}{Mean regression coefficient}

\item{ci_lower}{Lower CI bound}

\item{ci_upper}{upper confidence interval bound}
}
\value{
Data frame with pasted coefficients
}
\description{
Title Binds mean and 95% CI
}
