% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{calculate_crude_OR}
\alias{calculate_crude_OR}
\title{Title Runs crude odds ratios for each predictor variable in one execution}
\usage{
calculate_crude_OR(data, response_var, predictor_vars)
}
\arguments{
\item{data}{Data.frame containing response variable and all predictor variables}

\item{response_var}{Response variable for glm regression}

\item{predictor_vars}{Predictor variables for glm variables}
}
\value{
data.frame with cude regression coefficients for all predictor variables involved
}
\description{
Title Runs crude odds ratios for each predictor variable in one execution
}
