% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{clean_glm_outputs}
\alias{clean_glm_outputs}
\title{Title}
\usage{
clean_glm_outputs(glm_object)
}
\arguments{
\item{glm_object}{glm object as outputed from glm call}
}
\value{
Data frame with clean variable names and coefficients
}
\description{
Title
}
