% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{clean_regression_variables}
\alias{clean_regression_variables}
\title{Title Cleans variable names from regression outputs}
\usage{
clean_regression_variables(tab)
}
\arguments{
\item{tab}{data.frame with messy variable names and regression coefficients}
}
\value{
data.frame with clean variable names and coefficients
}
\description{
Title Cleans variable names from regression outputs
}
