% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{create_table}
\alias{create_table}
\title{Title Create publication quality table}
\usage{
create_table(data)
}
\arguments{
\item{data}{data.frame for tabble creation}
}
\description{
Title Create publication quality table
}
