% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{create_weighted_prevalence}
\alias{create_weighted_prevalence}
\title{Title Calculate weighted and weighted prevalence}
\usage{
create_weighted_prevalence(survey_data, survey_weight, group, response_var)
}
\arguments{
\item{survey_data}{Dataframe containing variables,survey weight etc}

\item{survey_weight}{Column in  dataframe for survey weight}

\item{group}{Column in the dataframe for grouping variables such as sex}

\item{response_var}{Column for a response variable for which prevalence is generated}
}
\description{
Title Calculate weighted and weighted prevalence
}
