% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepare_crude_regression}
\alias{prepare_crude_regression}
\title{Title This fits glm and outputs Crude OR for for each variable}
\usage{
prepare_crude_regression(data, response_var, predictor_vars)
}
\arguments{
\item{data}{Data.frame containing response variable and predictors}

\item{predictor_vars}{}
}
\value{
Data.frame containing processed coeffcients and variables
}
\description{
Title This fits glm and outputs Crude OR for for each variable
}
