% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{prepare_regression_output}
\alias{prepare_regression_output}
\title{Title This prepares regression outputs}
\usage{
prepare_regression_output(glm_object)
}
\arguments{
\item{glm_object}{This is  glm or lm class as outputed by glm}
}
\value{
Clean data.frame with for table
}
\description{
Title This prepares regression outputs
}
